basic input handling and broadcasting

+++ common/pkg/vec2/vec2.go
@@ -3,11 +3,11 @@ package vec2
 import "math"
 
 type Vec2 struct {
-	X float32
-	Y float32
+	X float64
+	Y float64
 }
 
-func NewVec2(x float32, y float32) Vec2 {
+func NewVec2(x float64, y float64) Vec2 {
 	return Vec2{x, y}
 }
 
@@ -21,20 +21,20 @@ func (v *Vec2) Sub(v2 Vec2) {
 	v.Y -= v2.Y
 }
 
-func (v *Vec2) Mul(s float32) {
+func (v *Vec2) Mul(s float64) {
 	v.X *= s
 	v.Y *= s
 }
 
-func (v *Vec2) Div(s float32) {
+func (v *Vec2) Div(s float64) {
 	v.X /= s
 	v.Y /= s
 }
 
-func (v *Vec2) Length() float32 {
-	return float32(math.Sqrt(float64(v.X*v.X + v.Y*v.Y)))
+func (v *Vec2) Length() float64 {
+	return float64(math.Sqrt(float64(v.X*v.X + v.Y*v.Y)))
 }
 
-func (v *Vec2) SqrLength() float32 {
+func (v *Vec2) SqrLength() float64 {
 	return v.X*v.X + v.Y*v.Y
 }
